Initial commit with old code
This is an old script I made over a year ago. Created this repo
as a way to keep track of it and as an excuse to finally get
around to re-writing it. Already made some minor formatting
changes so I'll cringe a little bit less pushing it, but it
still needs a lot of reworks.

.gitignore
* added JetBrains stuff
* added video file extensions

avisynthWriter.py
* old script I wrote a long time ago which is the motivation
behind this repo